+++ classes/archivingmod.php
@@ -47,6 +47,9 @@ class archivingmod extends \local_archiving\driver\archivingmod {
     /** @var int ID of the targeted quiz */
     protected int $quizid;
 
+    /** @var string Short-name of the web service the external worker uses for communication with Moodle */
+    public const WEB_SERVICE_SHORTNAME = 'archivingmod_quiz_ws';
+
     /**
      * Creates a new activity archiving driver instance.
      *
@@ -59,10 +62,10 @@ class archivingmod extends \local_archiving\driver\archivingmod {
         // Try to get course, cm info, and quiz.
         [$this->course, $this->cm] = get_course_and_cm_from_cmid($this->cmid, 'quiz');
         if (empty($this->cm)) {
-            throw new \moodle_exception('invalid_cmid', 'archivingmod_quiz');
+            throw new \moodle_exception('invalid_cmid', 'archivingmod_quiz'); // @codeCoverageIgnore
         }
         if ($this->course->id != $this->courseid) {
-            throw new \moodle_exception('invalid_courseid', 'archivingmod_quiz');
+            throw new \moodle_exception('invalid_courseid', 'archivingmod_quiz'); // @codeCoverageIgnore
         }
         $this->quizid = $this->cm->instance;
     }
@@ -72,14 +75,15 @@ class archivingmod extends \local_archiving\driver\archivingmod {
         $config = get_config('archivingmod_quiz');
 
         if (
-            intval($config->webservice_id ?? 0) <= 0 ||
-            intval($config->webservice_userid ?? 0) <= 0 ||
-            strlen($config->worker_url ?? '') < 1
+            strlen($config->worker_url ?? '') > 1 &&
+            self::get_webserviceid() > 0 &&
+            self::is_webservices_enabled() &&
+            self::is_webserviceproto_rest_enabled()
         ) {
-            return false;
-        } else {
             return true;
         }
+
+        return false;
     }
 
 
@@ -107,7 +111,7 @@ class archivingmod extends \local_archiving\driver\archivingmod {
 
     #[\Override]
     public function get_job_create_form(string $handler, \cm_info $cminfo): \local_archiving\form\job_create_form {
-        return new form\job_create_form($handler, $cminfo);
+        return new form\job_create_form($handler, $cminfo); // @codeCoverageIgnore
     }
 
     #[\Override]
@@ -120,9 +124,10 @@ class archivingmod extends \local_archiving\driver\archivingmod {
             // Prepare access to quiz and webservice.
             $quizmanager = quiz_manager::from_context($task->get_context());
             $attempts = $quizmanager->get_attempts();
+
             $wstoken = $task->create_webservice_token(
-                webserviceid: get_config('archivingmod_quiz', 'webservice_id'),
-                userid: get_config('archivingmod_quiz', 'webservice_userid'),
+                webserviceid: self::get_webserviceid(),
+                userid: get_admin()->id,
                 lifetimesec: get_config('local_archiving', 'job_timeout_min') * MINSECS
             );
 
@@ -138,6 +143,14 @@ class archivingmod extends \local_archiving\driver\archivingmod {
             $job->set_metadata_entry('num_attempts', count($attempts));
             $job->set_metadata_entry('num_attachments', $numattachments);
 
+            // Do not actually call the remote archive worker during unit tests.
+            if (defined('PHPUNIT_TEST') && PHPUNIT_TEST === true) {
+                $task->set_status(activity_archiving_task_status::AWAITING_PROCESSING);
+                throw new yield_exception();
+            }
+
+            // @codeCoverageIgnoreStart
+
             // Enqueue a new job at the worker.
             $worker = remote_archive_worker::instance();
             $workerjob = $worker->enqueue_archive_job(
@@ -150,6 +163,8 @@ class archivingmod extends \local_archiving\driver\archivingmod {
             // TODO (MDL-0): Error handling.
             $task->set_status(activity_archiving_task_status::AWAITING_PROCESSING);
             throw new yield_exception();
+
+            // @codeCoverageIgnoreEnd
         }
 
         if ($task->get_status(usecached: true) == activity_archiving_task_status::AWAITING_PROCESSING) {
@@ -223,5 +238,43 @@ class archivingmod extends \local_archiving\driver\archivingmod {
             'quiztimemodified' => $quiztimemodified,
             'attempttimemodified' => $attempttimemodified,
         ]);
+    }
+
+    /**
+     * Retrieves the ID of the self::WEB_SERVICE_SHORTNAME web service
+     *
+     * @return int ID of the web service
+     * @throws \dml_exception If the web service could not be found (should never happen,
+     * but who knows all the weird stats a Moodle instance can be in ...)
+     */
+    public static function get_webserviceid(): int {
+        global $DB;
+
+        return $DB->get_field(
+            'external_services',
+            'id',
+            ['shortname' => self::WEB_SERVICE_SHORTNAME],
+            MUST_EXIST
+        );
+    }
+
+    /**
+     * Determines if web services are enabled globally.
+     *
+     * @return bool True, if web services are enabled, false otherwise
+     * @throws \dml_exception
+     */
+    public static function is_webservices_enabled(): bool {
+        return get_config('core', 'enablewebservices') == true;
+    }
+
+    /**
+     * Determines if the web service protocol "REST" is enabled globally.
+     *
+     * @return bool True, if REST protocol is enabled, false otherwise
+     * @throws \dml_exception
+     */
+    public static function is_webserviceproto_rest_enabled(): bool {
+        return stripos(get_config('core', 'webserviceprotocols'), 'rest') !== false;
     }
 }