+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## Version X.Y.Z (YYYYMMDDXX)
 
+- Simplify web service setup process
+    - Bundle web service functions for worker communication inside a statically provided web service
+    - Remove superfluous admin settings for manual web service setups
+    - Remove superfluous autoinstall feature that was superseded by the statically provided web service
 - Finalize Moodle privacy API provider
+- Adapt web service unit tests to latest activity archiving task access token invalidation behavior
+- Create unit tests for activity archiving driver base class
 
 
 ## Version 0.3.0 (2025101300)